// backend/src/database/entities/enums.ts
export enum UserStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  SUSPENDED = 'SUSPENDED',
  PENDING = 'PENDING',
}

export enum MitraStatus {
  PENDING = 'PENDING',
  ACTIVE = 'ACTIVE',
  SUSPENDED = 'SUSPENDED',
}

export enum GeneralStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
}

export enum ProductType {
  MAIN_PRODUCT = 'MAIN_PRODUCT',
  ACCESSORY = 'ACCESSORY',
  MATERIAL = 'MATERIAL',
  SPAREPART = 'SPAREPART',
}

export enum OrderStatus {
  WAITING_CONFIRMATION = 'WAITING_CONFIRMATION',
  ASSIGNED = 'ASSIGNED',
  ACCEPTED = 'ACCEPTED',
  ON_THE_WAY = 'ON_THE_WAY',
  ARRIVED = 'ARRIVED',
  IN_PROGRESS = 'IN_PROGRESS',
  WAITING_APPROVAL = 'WAITING_APPROVAL',
  COMPLETED = 'COMPLETED',
  WAITING_PAYMENT = 'WAITING_PAYMENT',
  PAID = 'PAID',
  CANCELLED = 'CANCELLED',
}

export enum OrderItemType {
  PRODUCT = 'PRODUCT',
  ACCESSORY = 'ACCESSORY',
  SERVICE = 'SERVICE',
  ADDON = 'ADDON',
}

export enum AssignmentStatus {
  PENDING = 'PENDING',
  ACCEPTED = 'ACCEPTED',
  REJECTED = 'REJECTED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
}

export enum ChangeRequestStatus {
  WAITING = 'WAITING',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
}

export enum PaymentStatus {
  UNPAID = 'UNPAID',
  WAITING_PAYMENT = 'WAITING_PAYMENT',
  WAITING_PAYMENT_VERIFICATION = 'WAITING_PAYMENT_VERIFICATION',
  WAITING_CASH_COLLECTION = 'WAITING_CASH_COLLECTION',
  WAITING_ADMIN_CONFIRMATION = 'WAITING_ADMIN_CONFIRMATION',
  PAID = 'PAID',
  REJECTED = 'REJECTED',
  REFUNDED = 'REFUNDED',
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED',
}

export enum DiscountType {
  PERCENTAGE = 'PERCENTAGE',
  FIXED = 'FIXED',
}

export enum TargetUser {
  NEW_USER = 'NEW_USER',
  EXISTING_USER = 'EXISTING_USER',
  ALL_USER = 'ALL_USER',
}

export enum PromotionStatus {
  ACTIVE = 'ACTIVE',
  INACTIVE = 'INACTIVE',
  EXPIRED = 'EXPIRED',
}

export enum NotificationChannel {
  WHATSAPP = 'WHATSAPP',
  PUSH_NOTIFICATION = 'PUSH_NOTIFICATION',
  SMS = 'SMS',
  EMAIL = 'EMAIL',
}

export enum NotificationStatus {
  QUEUED = 'QUEUED',
  SENT = 'SENT',
  FAILED = 'FAILED',
}